ImageCaloriesCalc/requirements.txt
requests
beautifulsoup4
streamlit
numpy
Pillow
flask
keras~=2.12.0
tensorflow~=2.12.0